chore: add start command
Only allows read access to target dir and own dir.

Still open to abuse (like reading our .env file) if user starts probing
around and finds this source dir. For example, by specifying a path
like:

    ../../home/jm/log-creep/

I choose to treat the user as non-hostile and competent for the moment,
but note this as an area that needs to be addressed.