fix: replace regex with substring match
Simpler and less abuse/bug prone.

While here, refactors the name 'pattern' to 'match' to be more accurate.
Okay, so it's not a real refactor, since it changes the API, but I'm the
only one using this so there. :P

Also, I wish I could come up with a better name than 'match'.