fix: use ReadableStream instead of readline
More appropriate for chunking off pieces to read from the bottom of the
file first.

NOTE: This changeset is not complete - currently it buffers the entire
file (in chunks, but not in the order we want), so it is not optimized
for finding "new" lines (text at the bottom of the file).