wip: attempt to use Readable stream
I don't think I implemented this correctly, though. There aren't a lot
of good examples of this to draw from, and I think I might be
instantiating these ReadStream in a wrongheaded manner. I've also
noticed what appears to be chunks coming in out-of-order in my 'data'
handler.

I was trying to "section off" chunks of the file's buffer from the end of
the file, and then iterating backwards.

Going to revert this, but leave it in the history for elucidation.