fix: read in variable length buffers
Somehow NodeJS is fine without this - it just fills the overflow in the
buffer with a bunch of \x00 (hex for 0). Anyway, now the last buffer's
length with only be as long as it needs to work with the last N bytes.

Also uses 'text' variable to hang on to the line(s) we're processing at
any given iteration.

One potential improvement would be to leave all this as Buffer objects,
and use the .includes, .indexOf, and .slice methods to do all our
comparison and, and Buffer.concat to keep track of our current 'line'.