fix: don't crash app on ENOENT errors
Throw a "custom" Error object with a new statusCode property to
accomplish this.

Actually I think callbacks or event handlers might be better for this,
rather than twiddling with res.

On that note, this commit also removes a superfluous res.end() on empty
files, and just sets the status code and returns instead.

Resolves #3.