refactor: use method with positional arguments
This is required to make use of the 'memfs' virtual filesystem package
in our tests. Sadly, 'mock-fs' does not support promises, and memfs does
not actually mock every function and method of 'fs' (like for instance
the `filehandle.read(buffer[, options])` version of this method).

Honestly, this is just as clear as the version with named parameters in
our case.